Add way to get most probable bitstring
Removed the binary count squash mechanism as it's not used and not really how results are dealt with and replaced it with what the same result should have been used for. Left in binary from a QPU because it can still have use in niche situations. One of those things that is greatly useful if you need it.