ConvertToPydanticIR pass. (#251)
Framework and initial pass implementation that convert legacy IR and elements to their Pydantic counterparts.
I have opted for putting this in it's own module in qat.ir as it is not part of any particular end of the flow and is intended to move forward from the back as we implement more and more passes in Pydantic.

GitOrigin-RevId: 3edc8643db52b71491903103cb7c68c1c3a70274